+++ .github/workflows/test.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        go-version: ['1.22.x', '1.23.x', '1.24.x']
+        go-version: ['1.23.x', '1.24.x']
     steps:
       - uses: actions/checkout@v4
         with:
