+++ field/goldilocks/goldilocks_gnark.go
@@ -10,8 +10,6 @@ import (
 
 type Element = g.Element
 
-const Bytes = 8
-
 func NewElement(value uint64) Element {
 	return g.NewElement(value)
 }
@@ -88,13 +86,6 @@ func ToString(e ...Element) string {
 	return res
 }
 
-func FromBool(value bool) Element {
-	if value {
-		return One()
-	}
-	return Zero()
-}
-
 func FromInt64Abs(value int64) Element {
 	return FromUint64(uint64(value & 0x7FFFFFFFFFFFFFFF))
 }
@@ -117,10 +108,6 @@ func FromUint32(value uint32) Element {
 
 func Equals(a, b *Element) bool {
 	return a.Equal(b)
-}
-
-func Modulus() uint64 {
-	return g.Modulus().Uint64()
 }
 
 func Zero() Element {
@@ -148,39 +135,9 @@ func Sample() Element {
 	return elem
 }
 
-func RandArray(count int) []Element {
-	ret := make([]Element, count)
-	for i := 0; i < count; i++ {
-		ret[i] = Sample()
-	}
-	return ret
-}
-
-func Add(elems ...Element) Element {
-	res := g.NewElement(0)
-	for _, elem := range elems {
-		res.Add(&res, &elem)
-	}
-	return res
-}
-
-func Sub(a, b *Element) Element {
-	res := g.NewElement(0)
-	res.Sub(a, b)
-	return res
-}
-
-func Mul(elems ...*Element) Element {
-	res := g.NewElement(1)
-	for _, elem := range elems {
-		res.Mul(&res, elem)
-	}
-	return res
-}
-
 func Sqrt(elem *Element) *Element {
-	elemCopy := DeepCopy(elem)
-	return elemCopy.Sqrt(&elemCopy)
+	elemCopy := new(Element).Set(elem)
+	return elemCopy.Sqrt(elemCopy)
 }
 
 // Powers starting from 1
@@ -191,8 +148,4 @@ func Powers(e *Element, count int) []Element {
 		ret[i].Mul(&ret[i-1], e)
 	}
 	return ret
-}
-
-func DeepCopy(source *Element) Element {
-	return Element{source[0]}
 }
