+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/elliottech/poseidon_crypto
 
-go 1.22
+go 1.23
 
 require (
 	github.com/consensys/gnark-crypto v0.12.2-0.20240215234832-d72fcb379d3e
